requirements.txt
customtkinter
mss
numpy
imageio[ffmpeg]
pillow
screeninfo
pyaudio
soundcard
